--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,6 +16,7 @@ async-trait = "0.1"
 console = "0.15"
 futures-util = { version = "0.3", default-features = false, features = ["alloc"] }
 glob = "0.3"
+itertools = "0.13"
 libtest-mimic = "0.7"
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.9"