--- a/naivedb/Cargo.toml
+++ b/naivedb/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 [dependencies]
 anyhow = { version = "1", features = ["backtrace"] }
 async-trait = "0.1"
+clap = { version = "4.5", features = ["derive"] }
 sqlplannertest = { path = ".." }
 tokio = { version = "1", features = ["rt", "rt-multi-thread", "macros", "fs"] }