--- a/naivedb/tests/basics/create.planner.sql
+++ b/naivedb/tests/basics/create.planner.sql
@@ -1,0 +1,7 @@
+-- Test whether select stmt works correctly.
+CREATE TABLE t1(v1 int);
+
+/*
+
+*/
+