--- a/naivedb/tests/basics/create.planner.sql
+++ b/naivedb/tests/basics/create.planner.sql
@@ -2,6 +2,13 @@
 CREATE TABLE t1(v1 int);
 
 /*
+=== logical
+I'm a naive db, so I don't now how to process
+CREATE TABLE t1(v1 int);
 
+
+=== physical
+I'm a naive db, so I don't now how to process
+CREATE TABLE t1(v1 int);
 */