--- a/naivedb/tests/basics/create.yml
+++ b/naivedb/tests/basics/create.yml
@@ -1,4 +1,6 @@
 - sql: |
     CREATE TABLE t1(v1 int);
   desc: Test whether select stmt works correctly.
-  
+  tasks:
+  - logical
+  - physical