--- a/naivedb/tests/basics/select.planner.sql
+++ b/naivedb/tests/basics/select.planner.sql
@@ -1,0 +1,7 @@
+-- Test whether select stmt works correctly.
+SELECT * FROM t1;
+
+/*
+
+*/
+