--- a/naivedb/tests/basics/select.planner.sql
+++ b/naivedb/tests/basics/select.planner.sql
@@ -2,6 +2,15 @@
 SELECT * FROM t1;
 
 /*
+=== logical
+I'm a naive db, so I don't now how to process
+CREATE TABLE t1(v1 int);
+SELECT * FROM t1;
 
+
+=== physical
+I'm a naive db, so I don't now how to process
+CREATE TABLE t1(v1 int);
+SELECT * FROM t1;
 */