--- a/naivedb/tests/basics/select.yml
+++ b/naivedb/tests/basics/select.yml
@@ -1,0 +1,4 @@
+- sql: |
+    SELECT * FROM t1;
+  desc: Test whether select stmt works correctly.
+  before: ["CREATE TABLE t1(v1 int);"]