--- a/naivedb/tests/basics/select.yml
+++ b/naivedb/tests/basics/select.yml
@@ -2,4 +2,6 @@
     SELECT * FROM t1;
   desc: Test whether select stmt works correctly.
   before: ["CREATE TABLE t1(v1 int);"]
-  
+  tasks:
+    - logical
+    - physical