Adding the ability to match selected test modules or files. Also display test names with module hierarchy.

- can match globs in planner test apply.
- show module hierarchy in both apply + tests.

**Before**

<pre>
$ cargo run -p naivedb --bin apply
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.03s
     Running `target/debug/apply`
[DONE] ok.yml, took 1 ms
[DONE] error.yml, took 1 ms

$ cargo nextest run -p naivedb ok
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.02s
------------
 Nextest run ID 44de9a31-bc77-41c3-90d5-def47c4694dd with nextest profile: default
    Starting 1 test across 3 binaries (1 test skipped)
        PASS [   0.005s] naivedb::plannertest ok
------------
     Summary [   0.005s] 1 test run: 1 passed, 1 skipped
</pre>

**After**
<pre>
$ cargo run -p naivedb --bin apply -- basics ok
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.09s
     Running `target/debug/apply basics ok`
<b>[DONE] basics/create.yml, took 1 ms</b>
<b>[DONE] basics/select.yml, took 1 ms</b>
[DONE] ok.yml, took 1 ms

$ cargo nextest run -p naivedb <b>basics::create ok</b>
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.02s
------------
 Nextest run ID 3789c774-11bc-4d06-98ec-2d8dde74789e with nextest profile: default
    Starting 2 tests across 3 binaries (2 tests skipped)
        PASS [   0.005s] naivedb::plannertest <b>basics::create</b>
        PASS [   0.004s] naivedb::plannertest ok
------------
     Summary [   0.005s] 2 tests run: 2 passed, 2 skipped
</pre>